refactor(CreateRandom): 첫화면, 안내사항 컴포넌트 분리
